FINAL STABLE
In this update, I added feature to let cHash deletes all the elements in its internal array. The function to perform this task is 'Sweep()' which attached as a member method to cHash.

NB: 'Sweep()' is only wiped-off all the data elements in the array, not the array itself. so, the array can still maintain the previous allocated spaces in memory and the size or number of block of spaces are kept in maintained as well.