FINAL STABLE
2024/OCT/02:
Updated 'iterRight()' and 'enforce_unique()' functions.
the current update subsequently add one to each bit of the right node that has a tendency of duplicating the same bits pattern so that every created bits pattern is now independently unique to each other. These functions are now migrated to private scope of the _TREE functional structure.